requirements.txt
alembic==1.0.8
amqp==2.4.2
attrs==19.1.0
billiard==3.6.0.0
blinker==1.4
boto3==1.9.130
botocore==1.12.130
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
docutils==0.14
flasgger==0.9.2
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Migrate==2.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.4
jsonschema==3.0.1
kombu==4.5.0
Mako==1.0.7
MarkupSafe==1.1.1
mistune==0.8.4
psycopg2==2.7.7
PyJWT==1.7.1
pyrsistent==0.14.11
PySocks==1.6.8
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2018.9
PyYAML==5.1
raven==6.10.0
redis==3.2.1
requests==2.21.0
s3transfer==0.2.0
sentry-sdk==0.7.10
six==1.12.0
SQLAlchemy==1.3.0
SQLAlchemy-Utils==0.33.11
twilio==6.26.0
urllib3==1.24.1
vine==1.3.0
voluptuous==0.11.5
Werkzeug==0.14.1
Flask-JWT-Extended==3.3.4
docxtpl==0.5.17
flask-uploads==0.2.1